Modified project to build jar
