+++ sop/build.gradle
@@ -1,0 +1,1 @@
+version = '1.0'
